requirements.txt
discord.py==1.5.1
python-dotenv==0.15.0
spotipy==2.16.1
SQLAlchemy==1.3.20
psycopg2-binary==2.8.6
scipy==1.6.2
numpy==1.20.2
pandas==1.2.4
